+++ packages/ui/CHANGELOG.md
@@ -6,6 +6,7 @@ The following is a curated list of changes in the Enact ui module, newest change
 
 ### Added
 
+- `ui/RadioDecorator` and `ui/RadioControllerDecorator` to support radio group-style management of components
 - `onAppear`, `onEnter`, `onLeave`, `onStay`, `onTransition`, and `onWillTransition` events to ViewManager
 - `none` to `scrimType` prop values in `ui/FloatingLayer`